Implemented add image feature for businesses
At the registration step, businesses are required to add an image to be associated with their business.
The images are stored in the firebase storage, under "business banners" folder.
The images are named  "< business uid>_banner.<file extension>".
Each time an image is added, an "image" class containing the image name and image download url is added under the registered business' database.
This can be used to obtain the image to display it along with its respective business
Things to check before merge:
- The business class, might need to have a "businessImage" field. Discussion needed.